Move annotator.plugin.highlighter -> annotator.ui.highlighter.standalone
